pyproject.toml
[project]
name = "digital-amber"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "mammoth>=1.11.0",
    "python-docx>=1.2.0",
    "markdown>=3.5.0",
    "weasyprint>=60.0",
    "ebooklib>=0.18",
    "replicate>=0.25.0",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "pyyaml>=6.0.3",
]
